concepts: make Orientation::length() unsigned, enforce it in Tweenable (finding 10)
Orientation::length() returned int (the class's internal index idiom) while
every other Tweenable count is size_t. A signed count could wrap negative into
a huge loop bound, leaving non-negativity to convention. Return size_t and
tighten the Tweenable concept from convertible_to<size_t> to unsigned_integral
so the guarantee is type-enforced.

+++ core/geometry.h
@@ -586,7 +586,7 @@ public:
    * @brief Gets the number of recorded orientation frames in the current step.
    * @return The number of frames.
    */
-  int length() const { return num_frames; }
+  size_t length() const { return static_cast<size_t>(num_frames); }
 
   /**
    * @brief Rotates a vector by the current (latest) orientation.
